timer: move utils/AppTimer to services/TimerService
